requirements.txt
# Nerpa GUI Requirements
# Основные зависимости для графического интерфейса Nerpa

# Streamlit - фреймворк веб-приложения
streamlit>=1.28.0

# Обработка данных
pandas>=2.0.0
numpy>=1.24.0

# Визуализация
plotly>=5.17.0

# Биоинформатика
biopython>=1.81

# Системные утилиты (встроенные в Python 3.7+, но указываем для ясности)
# pathlib - работа с путями (встроен)
# typing - type hints (встроен)
# dataclasses - структуры данных (встроен в Python 3.7+)

# Для работы с логами
# logging - встроен

# Опциональные зависимости для улучшения функционала
typing-extensions>=4.5.0  # Расширенная поддержка type hints

# Для production deployment (опционально)
# streamlit-autorefresh>=0.0.1  # Автообновление
# streamlit-extras>=0.3.0  # Дополнительные компоненты
